+++ packages/sources/coinmetrics-test/src/endpoint/price-ws.ts
@@ -111,7 +111,7 @@ export const handleAssetMetricsMessage = (
               result,
             },
             timestamps: {
-              providerIndicatedTime: new Date(message.time).getTime(),
+              providerIndicatedTimeUnixMs: new Date(message.time).getTime(),
             },
           },
         }