+++ packages/sources/elwood/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/elwood-adapter
 
+## 2.0.1
+
+### Patch Changes
+
+- 59c4a2ee4: Bumped v3 framework version
+
 ## 2.0.0
 
 ### Major Changes