+++ packages/sources/kaiko-test/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/kaiko-test-adapter
 
+## 1.0.2
+
+### Patch Changes
+
+- c152ed06e: Bumped v3 framework version
+
+## 1.0.1
+
+### Patch Changes
+
+- fa05cc85b: Removed rate limit config
+
 ## 1.0.0
 
 ### Major Changes